+++ src/pages/copyandedit/CopyAndEditPaste.tsx
@@ -2,12 +2,14 @@ import { Paste } from "pasty-core";
 import * as React from "react";
 import { connect, Dispatch } from "react-redux";
 
+import { getPaste } from "actions/creators";
+import config from "configfile";
 import { customInitialCreatePaste } from "pages/paste/CreatePaste";
-import ViewPaste from "pages/view/ViewPaste";
 
+import GenericNonIdealState from "components/GenericNonIdealState";
+import PasteLoading from "components/PasteLoading";
 import Maybe from "helpers/maybe";
-import { STATE as CREATE_STATE } from "pages/paste/reducer";
-import { STATE as VIEW_STATE } from "pages/view/reducer";
+import { STATE } from "pages/view/reducer";
 import { PasteFileTypes } from "reducers/form";
 import { IReducer } from "reducers/index";
 
@@ -25,21 +27,78 @@ export interface ICopyAndEditPasteProps {
   };
 }
 
-export interface ICopyAndEditPasteDispatchProps {
-  viewState: VIEW_STATE;
+export interface ICopyAndEditPasteStateProps {
+  state: STATE;
   paste: Maybe<Paste>;
 }
 
-export type PropsType = ICopyAndEditPasteDispatchProps & ICopyAndEditPasteProps;
+export interface ICopyAndEditPasteDispatchProps {
+  getPasteAction: (id: string, key: string, highlight: number[][]) => void;
+}
+
+export interface ICopyAndEditPasteState {
+  error?: Error;
+  errorInfo?: React.ErrorInfo;
+}
+
+export type PropsType = ICopyAndEditPasteDispatchProps & ICopyAndEditPasteStateProps & ICopyAndEditPasteProps;
 
 /*
-   Loads a remote paste from url parameters and then allows
-   editing. Hijacks a ViewPaste component and hides to load the
-   paste. This is likely fragile.
+  Behaves similarly to ViewPaste except that, when it's ready to view, it creates
+  a custom CreatePaste form.
 */
-export class CopyAndEditPaste extends React.PureComponent<PropsType> {
+export class CopyAndEditPaste extends React.Component<PropsType, ICopyAndEditPasteState> {
+  public constructor(props) {
+    super(props);
+
+    this.state = {};
+  }
+
+  public componentWillMount() {
+    if (!this.props.paste.isNothing()
+        && this.props.paste.getData().name === this.props.match.params.id
+        && this.props.paste.getData().key === this.props.match.params.key) {
+      return;
+    }
+
+    this.props.getPasteAction(
+      this.props.match.params.id,
+      this.props.match.params.key,
+      [],
+    );
+  }
+
+  public componentDidCatch(error: Error, errorInfo: React.ErrorInfo) {
+    this.setState({
+      error,
+      errorInfo,
+    });
+  }
+
   public render() {
     let files = {};
+
+    if (this.props.state !== STATE.VIEWING) {
+      return (
+        <PasteLoading
+          firstStageLabel="Downloaded"
+          secondStageLabel="Decrypted"
+          firstStageKey={STATE.DOWNLOADING}
+          secondStageKey={STATE.DECRYPTING}
+          type="viewPaste"
+        />
+      );
+    }
+
+    if (this.state.error) {
+      return (
+        <GenericNonIdealState
+          error={this.state.error}
+          errorInfo={this.state.errorInfo}
+          action="copying and editing a paste"
+        />
+      );
+    }
 
     if (!this.props.paste.isNothing()) {
       files = {
@@ -50,24 +109,22 @@ export class CopyAndEditPaste extends React.PureComponent<PropsType> {
       };
     }
 
-
     return (
-      <React.Fragment>
-        <ViewPaste
-          hidden={true}
-          location={this.props.location}
-          match={this.props.match}
-        />
-        {this.props.viewState === VIEW_STATE.VIEWING && React.createElement(customInitialCreatePaste(files))}
-      </React.Fragment>
+      this.props.state === STATE.VIEWING && React.createElement(customInitialCreatePaste(files))
     );
   }
 }
 
 export const mapStateToProps = (state: IReducer, ownProps: ICopyAndEditPasteProps)
-                             : ICopyAndEditPasteDispatchProps => ({
+                             : ICopyAndEditPasteStateProps => ({
   paste: state.viewPaste.paste,
-  viewState: state.viewPaste.state,
+  state: state.viewPaste.state,
 });
 
-export default connect(mapStateToProps, {})(CopyAndEditPaste);
+export const mapDispatchToProps = (dispatch: Dispatch<IReducer>): ICopyAndEditPasteDispatchProps => ({
+  getPasteAction: (id: string, key: string) => (
+    dispatch(getPaste(id, key, [], `${config.get}${id}`))
+  ),
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(CopyAndEditPaste);